@AuthenticationPrincipal UserDeatils 적용 - 댓글 기능에서 사용자 정보 받기

+++ src/main/java/com/sparta/schedulemanagement_spring/service/CommentsService.java
@@ -4,6 +4,7 @@ import com.sparta.schedulemanagement_spring.dto.CommentsRequestDto;
 import com.sparta.schedulemanagement_spring.dto.CommentsResponseDto;
 import com.sparta.schedulemanagement_spring.entity.Comments;
 import com.sparta.schedulemanagement_spring.entity.Schedule;
+import com.sparta.schedulemanagement_spring.entity.User;
 import com.sparta.schedulemanagement_spring.repository.CommentsRepository;
 import com.sparta.schedulemanagement_spring.repository.ScheduleRepository;
 import lombok.RequiredArgsConstructor;
@@ -12,7 +13,6 @@ import org.springframework.stereotype.Service;
 
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
@@ -20,11 +20,11 @@ public class CommentsService {
     private final ScheduleRepository scheduleRepository;
     private final CommentsRepository commentsRepository;
 
-    public CommentsResponseDto addComment(Long scheduleId, CommentsRequestDto requestDto) {
+    public CommentsResponseDto addComment(Long scheduleId, CommentsRequestDto requestDto, User user) {
         Schedule schedule = scheduleRepository.findById(scheduleId).orElseThrow(()->
                 new IllegalArgumentException("일정을 찾을 수 없습니다."));
 
-        Comments comment = new Comments(requestDto,schedule);
+        Comments comment = new Comments(requestDto,schedule,user);
         Comments saveComment = commentsRepository.save(comment);
         CommentsResponseDto responseDto = new CommentsResponseDto(saveComment);
 
@@ -44,12 +44,16 @@ public class CommentsService {
     }
 
 
-    public CommentsResponseDto updateComment(Long scheduleId, Long commentId, CommentsRequestDto requestDto) {
+    public CommentsResponseDto updateComment(Long scheduleId, Long commentId, CommentsRequestDto requestDto, User user) {
         Schedule schedule = scheduleRepository.findById(scheduleId).orElseThrow(()->
                 new IllegalArgumentException("일정을 찾을 수 없습니다."));
         // 일정이 존재할 경우에만
+        // 선택한 댓글의 작성자와 현재 로그인한 사용자가 같을 때만
             Comments comments = commentsRepository.findById(commentId).orElseThrow(() ->
                     new IllegalArgumentException("댓글을 찾을 수 없습니다."));
+            if(comments.getUser().getId() != user.getId()){
+                throw new IllegalArgumentException("댓글 작성자가 아니므로 수정할 수 없습니다.");
+            }
             // 댓글 내용만 수정
             comments.setComment(requestDto.getComment());
             // DB에 수정된 댓글 저장
@@ -57,18 +61,20 @@ public class CommentsService {
 
         return new CommentsResponseDto(updateComments);
     }
-    public String deleteComment(Long scheduleId, Long commentId) {
+    public String deleteComment(Long scheduleId, Long commentId, User user) {
         Schedule schedule = scheduleRepository.findById(scheduleId).orElseThrow(()->
                 new IllegalArgumentException("일정을 찾을 수 없습니다."));
         // 일정이 존재할 경우에만
+        // 선택한 댓글의 작성자와 현재 로그인한 사용자가 같을 때만
             Comments comments = commentsRepository.findById(commentId).orElseThrow(()->
                     new IllegalArgumentException("댓글을 찾을 수 없습니다."));
 
-            try {
-                commentsRepository.delete(comments);
-            }catch (EmptyResultDataAccessException e){
-                e.getMessage();
-            }
+        if(comments.getUser().getId() != user.getId()){
+            throw new IllegalArgumentException("댓글 작성자가 아니므로 삭제할 수 없습니다.");
+        }
+
+        commentsRepository.delete(comments);
+
 
         return "일정 삭제에 성공하였습니다.";
     }
